+++ CHANGELOG.rst
@@ -8,6 +8,22 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 UNRELEASED CHANGES
 ******************
 
+0.3.12
+******
+* Revert boogie image version to 0.2.7 due to a discovered breaking issue in 0.2.8
+
+0.3.11
+******
+* Fetch tango boogie image from harbor instead of gitlab + update version
+
+0.3.10
+******
+* CIP-5078: Enable HealthState for VCC All Bands Simulation
+
+0.3.9
+*****
+* CIP-5528: Moved LRC results buffer into base simulator class
+
 0.3.8
 *****
 * CIP-5784: Add new transitions for end scan and scan command failures to go back to the previous state instead of incorrectly transitioning to the next state